Update to Poll and List
added a getID and getContent to both poll and list.  getContent for poll will return a list of poll_elements. poll_element has a getItem and getVotes. i can change this to return a string containing both values, or switch to a Map, with the key being votes (will take time to implement)